requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r prod.txt

# Deployment
Fabric

# Testing
nose
selenium
webtest
django-webtest
django-nose
coverage
django-coverage
factory_boy
watchdog

# Shell
ipython

# Command extensions
django-extensions
